Initial feature including commit of the database package.
The functionality is based on the idea of wrapping the creation of data into BaseTable.java child-classes, which all contain an instance of the tables Datasource.

The next commit will focus on completion of the BaseTable.java functionality, along with extensive unit testing of current features.

After this, functionality for retrieving configuration file data, for the purposes of getting data source information, will be developed.

Due to the nature of this package, extensive unit testing will be required before a merge can be commited, and said tests will be added to the github repo to ensure future-proofing of the database package.